requirements.txt
pyserial==3.5
python-dotenv==1.0.1
python-socketio==5.11.4
requests==2.32.3
RPi.GPIO==0.7.1
